buoy_common: move configs from buoy_sim to buoy_common
