+++ crates/buoy_sim/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod logging;
+
+pub mod prelude {
+    pub use buoy_common::*;
+}
+
